Added ChartRender util, charts now displayed in SlideViewPanel and PresentationSlide
