+++ components/Deck/ContentPanel/SlideModes/SlideViewPanel/SlideViewPanel.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {NavLink} from 'fluxible-router';
 import {connectToStores} from 'fluxible-addons-react';
 import SlideViewStore from '../../../../../stores/SlideViewStore';
+import ChartRender from '../../../util/ChartRender';
 import ResizeAware from 'react-resize-aware';
 import { findDOMNode } from 'react-dom';
 const ReactDOM = require('react-dom');
@@ -57,72 +58,7 @@ class SlideViewPanel extends React.Component {
         );
     }
     componentDidMount(){
-      // Create the charts
-      $("div[id^=chart]").each(function(){
 
-         // Extract the data of the chartID
-         let chart = JSON.parse($(this).attr('datum'));
-         let chartID = chart.data['this.chartID'];
-         let chartType =  chart.data.chartType;
-         let chartData = chart.data.chartData;
-         let data = null;
-         chart = null;
-        switch (chartType) {
-            case 'lineChart':
-                data = chartData;
-                chart = nv.models.lineChart()
-                  .useInteractiveGuideline(true);
-                chart.xAxis.tickFormat(function(d) { return chartData[0].xlabels[d] || d; });
-                break;
-            case 'barChart':
-                data = chartData;
-                chart = nv.models.multiBarChart();
-                chart.xAxis.tickFormat(function(d) { return chartData[0].xlabels[d] || d; });
-                break;
-            case 'pieChart':
-            case 'pie3DChart':
-                data = chartData[0].values;
-                chart = nv.models.pieChart();
-                break;
-            case 'areaChart':
-                data = chartData;
-                chart = nv.models.stackedAreaChart()
-                  .clipEdge(true)
-                  .useInteractiveGuideline(true);
-                chart.xAxis.tickFormat(function(d) { return chartData[0].xlabels[d] || d; });
-                break;
-            case 'scatterChart':
-
-                for (let i=0; i<chartData.length; i++) {
-                    let arr = [];
-                    for (let j=0; j<chartData[i].length; j++) {
-                        arr.push({x: j, y: chartData[i][j]});
-                    }
-                    data.push({key: 'data' + (i + 1), values: arr});
-                }
-
-
-                chart = nv.models.scatterChart()
-                  .showDistX(true)
-                  .showDistY(true)
-                  .color(d3.scale.category10().range());
-                chart.xAxis.axisLabel('X').tickFormat(d3.format('.02f'));
-                chart.yAxis.axisLabel('Y').tickFormat(d3.format('.02f'));
-                    break;
-            default:
-        }
-
-        if (chart !== null) {
-
-          d3.select('#' + chartID)
-            .append('svg')
-            .datum(chartData)
-            .transition().duration(500)
-            .call(chart);
-
-          // nv.utils.windowResize(chart.update);
-        }
-      });
 
         if(process.env.BROWSER){
 
@@ -153,6 +89,7 @@ class SlideViewPanel extends React.Component {
     }
     componentDidUpdate() {
         this.resize();
+        ChartRender.createCharts();
     }
     resize()
     {
