+++ components/Deck/Presentation/PresentationSlide.js
@@ -4,13 +4,20 @@ import {NavLink} from 'fluxible-router';
 import {connectToStores, provideContext } from 'fluxible-addons-react';
 import PresentationStore from '../../../stores/PresentationStore';
 import SlideViewStore from '../../../stores/SlideViewStore';
+import ChartRender from '../util/ChartRender';
 
 let sectionStyle = { 'top': 'unset !important'};
 
 class PresentationSlide extends React.Component {
 		componentDidMount(){
 		//this.props.style = sectionStyle;
+			this.forceUpdate();
 		}
+
+		componentDidUpdate(){
+			ChartRender.createCharts();
+		}
+
 		render(){
 
 				return (
